test: 单元测试完善

+++ core/src/main/java/org/gaius/octopus/core/pojo/dto/DatasourceInterfaceDTO.java
@@ -16,12 +16,12 @@ public class DatasourceInterfaceDTO {
     /**
      * 接口ID
      */
-    private Integer id;
+    private Long id;
     
     /**
      * 数据源ID
      */
-    private Integer datasourceId;
+    private Long datasourceId;
     
     /**
      * 接口名称
