+++ core/src/main/resources/schema-mysql.sql
@@ -1,7 +1,7 @@
 -- 用户表
 CREATE TABLE `user`
 (
-    `id`       int(11)      NOT NULL AUTO_INCREMENT comment '主键',
+    `id`       bigint       NOT NULL AUTO_INCREMENT comment '主键',
     `username` varchar(64)  not null comment '用户名',
     `password` varchar(512) not null comment '密码',
     `enabled`  tinyint(4)   not null comment '是否启用',
@@ -13,7 +13,7 @@ CREATE TABLE `user`
 -- 数据源表
 CREATE TABLE `datasource`
 (
-    `id`          int(11)     NOT NULL AUTO_INCREMENT comment '主键',
+    `id`          bigint      NOT NULL AUTO_INCREMENT comment '主键',
     `name`        varchar(64) not null comment '名称',
     `type`        tinyint(4) comment '类型',
     `description` varchar(255) comment '描述',
@@ -31,7 +31,7 @@ CREATE TABLE `datasource`
 -- 数据源接口表
 CREATE TABLE `datasource_interface`
 (
-    `id`            int(11)     NOT NULL AUTO_INCREMENT comment '主键',
+    `id`            bigint      NOT NULL AUTO_INCREMENT comment '主键',
     `name`          varchar(64) not null comment '名称',
     `datasource_id` int         not null comment '数据源ID',
     `description`   varchar(255) comment '描述',
